+++ extensions/iceworks-time-master/src/typings/package.ts
@@ -1,3 +1,0 @@
-export interface IPackageJson {
-  name: string;
-}