+++ packages/transform-ts-to-js/src/babelDecoratorPlugin.ts
@@ -1,9 +1,0 @@
-import syntaxDecorators from '@babel/plugin-syntax-decorators';
-
-export default function () {
-  return {
-    name: 'syntax-decorators',
-    inherits: syntaxDecorators,
-    visitor: {},
-  };
-}