+++ packages/transform-ts-to-js/src/babelTransformTSPlugin.ts
@@ -1,9 +1,0 @@
-import TransformTypescript from '@babel/plugin-transform-typescript';
-
-export default function () {
-  return {
-    name: 'transform-typescript',
-    inherits: TransformTypescript,
-    visitor: {},
-  };
-}