+++ scripts/publish-beta-extension.ts
@@ -3,14 +3,14 @@
  */
 import * as path from 'path';
 import { spawnSync } from 'child_process';
-import uploadExtesions from './upload-extensions';
-import { IExtensionInfo, getExtensionInfos } from './getExtensionInfos';
+import uploadExtesions from './fn/upload-extensions';
+import { IExtensionInfo, getExtensionInfos } from './fn/getExtensionInfos';
 import sleep from './fn/sleep';
 import checkPackagePublished from './fn/checkPackagePublished';
 import extensionDepsInstall from './fn/extension-deps-install';
 import updateExtensionDependencies from './fn/updateExtensionDependencies';
 
-function publish(extension: string, directory: string, version: string): void {
+function packExtension(extension: string, directory: string, version: string): void {
   // vsce package
   console.log('[VSCE] PACKAGE: ', `${extension}@${version}`);
   spawnSync('vsce', ['package'], {
@@ -50,7 +50,7 @@ async function start() {
       publishedCount++;
       console.log(`--- ${name}@${localVersion} ---`);
 
-      publish(name, directory, localVersion);
+      packExtension(name, directory, localVersion);
       publishedExtensions.push(`${name}:${localVersion}`);
     }
     uploadExtesions(publishedExtensions);