+++ extensions/iceworks-material-helper/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+# 0.5.0
+
+- refactor: internal optimization
+
+## 0.4.2
+
+- fix: fail to clear material sources
+- fix: fail to debug material sources in some cases
+- feat: layout path is not the necessary option in page-generator extension
+
 ## 0.4.1
 
 - fix: ui build error