+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+# 0.4.0
+
+- refactor: internal optimization
+
+# 0.3.18
+
+- refactor: material source card
+
 # 0.3.17
 
 - fix: build error